Load Balancing for Recognizer Works (#6)
Adds a new 'LoadBalancer' Flow component, which manages
a pool of Works. The 'LoadBalancer' will start/stop the Works
as the load increases, and distributes calls their 'run()' method.

The API was designed with the developer experience in mind, with
the goal that the app developer can treat a dynamic group of Works
as just a single Work. Hopefully, this component can be published
as a standalone component for use in other apps.