Faster First Echo Request Processing
Adds logic to make a dummy 'recognizer.run()' call on app startup.
In the cloud, this means a Work machine will be spun up and
ready before the first actual 'create_echo' request comes in,
and will not be blocked by spinning up a new machine.

In the future, hopefully the Lightning Framework will support
a 'LightningWork.start()' method or something similar so we
can remove this hack.