Basic Multi-User Support
Adds a 'user_id' field to the Echo model for associating Echoes
with a specific user. Alongside this is added the 'login'
command and API, which is used by the CLI and UI clients to
receive a unique user ID generated by the server which they
will attach to all of their requests to the API.

When using the CLI, 'lightning login' must be ran before
any other commands. In the UI, the user ID will be fetched
on app startup without explicit user action and stored in
local storage.

Also adds 'ECHO_ENABLE_MULTI_TENANCY' environment variable.
This is false by default, which causes the app to issue every
login request the same default user ID (single user mode).
When enabled, the server will generated unique user IDs on every
login request.