Fix Recorded Audio Playback on iOS
Adds 'python-magic' library to correctly identify the mimetype
of the file in the download handler. Audio recorded in the
browser is 'video/webm' which is not supported in iOS, so
we convert it to 'audio/mp3' in the recognizer and persist
that file to the Drive instead.

This fixes a regression introduced when we were incorrectly
using the 'mimetypes' module, which only guesses based on
extension.