Implement REST API
Previously, we were relying on the basic HTTP API created as a
result of using 'configure_commands()', which worked. But it
is easy to use 'configure_api()' to define proper REST API routes
and just call the existing methods we were using before. The
advantage of using 'configure_api()' is that we can return proper
HTTP error codes (like HTTP 404).