Validation Endpoint to Prevent Invalid Echo Creation
Adds a new '/api/validate' endpoint which validates
all the fields on an Echo and returns whether it passed
validation.

This is needed because the Lightning Framework does not
support raising HTTP exceptions in the 'configure_api()'
handlers without causing the entire Flow to crash, which
prevents us from returning proper HTTP 400 error codes
to the client. Instead, we will use this validation
endpoint in the UI.

Also adds the 'ECHO_VIDEO_SOURCE_MAX_DURATION_SECONDS'
environment variable to configure a video length limit
when creating an Echo from a YouTube URL. The primary
motiviation for this is to enable scaling in a hosted
multi-tenant environment in the cloud. The default
limit is 15 minutes.