Enable Polling When Not Recording
We previously removed polling to reduce load on the API server
and fix a bug during recording caused by too many re-renders.
However, without polling the list of Echoes will not update
when an Echo completes processing, giving the impression that
the app is slower than it actually is. This adds back polling
but disables it if the microphone is recording using a new
<RecordEchoContextProvider> and associated hooks.