Return Proper HTTP Status Codes for Errors
With the latest release of the Lightning Framework, we can now
use 'raise HTTPException(status_code=4xx)' in the
'configure_api()' handlers to return proper HTTP status codes
to the browser and other REST API clients. In the previous
version of the framework, raising any exception in these handlers
would cause the entire Flow to fail and crash the app.