Use FileServer Instead of Drive in YouTuber Component
Deleting Echoes was not working because Lightning Drives only
allow the component which added a file to delete it. When
creating an Echo from a YouTube URL, the YouTuber component
was adding the file directly into the Drive, which prevented
the FileServer component from deleting it.

This changes the YouTuber component to instead upload the file
to the FileServer component instead of putting it directly into
the Drive.