Use C++ Whisper Implementation (#119)
Switches to use the C++ implementation of the Whisper model,
which is faster and is capable of running efficiently on just
CPUs. This will save a lot of money when running in the cloud
because we don't need to use expensive GPUs and can just use
cheap CPU-only machines. This model also performs much better
when running locally on your laptop, whereas before trying to
run Echo on any video >5 minutes would take forever.